# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TerrAPI is a service designed to simplify and automate your deployments.",
        "title": "TerrAPI Controller",
        "contact": {
            "name": "Support",
            "url": "https://github.com/terrapi-solution",
            "email": "contact@thomas-illiet.fr"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/modules": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                  "🍆 Module"
                ],
                "summary": "List all modules.",
                "parameters": [
                    {
                      "type": "string",
                      "description": "Search",
                      "name": "search",
                      "in": "query"
                    },
                  {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "collectionFormat": "csv",
                    "description": "Filter",
                    "name": "filter",
                    "in": "query"
                  },
                  {
                    "minimum": 1,
                        "type": "integer",
                        "default": 1,
                    "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                  {
                    "type": "string",
                    "default": "id",
                    "description": "Order by",
                    "name": "order_by",
                    "in": "query"
                  },
                  {
                    "type": "string",
                    "default": "desc",
                    "description": "Order direction",
                    "name": "order_direction",
                    "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                              "$ref": "#/definitions/modules.ModuleResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
          "post": {
            "consumes": [
              "application/json"
            ],
            "produces": [
              "application/json"
            ],
            "tags": [
              "🍆 Module"
            ],
            "summary": "Create a new module.",
            "parameters": [
              {
                "description": "Module",
                "name": "module",
                "in": "body",
                "required": true,
                "schema": {
                  "$ref": "#/definitions/modules.ModuleResponse"
                }
              }
            ],
            "responses": {
              "201": {
                "description": "Created",
                "schema": {
                  "$ref": "#/definitions/modules.ModuleResponse"
                }
              },
              "400": {
                "description": "Bad Request",
                "schema": {
                  "$ref": "#/definitions/errors.AppError"
                }
              },
              "500": {
                "description": "Internal Server Error",
                "schema": {
                  "$ref": "#/definitions/errors.AppError"
                }
              }
            }
          }
        },
      "/v1/modules/{id}": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "🍆 Module"
          ],
          "summary": "Get a specific module.",
          "parameters": [
            {
              "type": "integer",
              "description": "Module identifier",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/modules.ModuleResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            }
          }
        },
        "delete": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "🍆 Module"
          ],
          "summary": "Delete a specific module.",
          "parameters": [
            {
              "type": "integer",
              "description": "Module identifier",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            }
          }
        }
      },
      "/v1/modules/{id}/source": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "🍆 Module"
          ],
          "summary": "Get the source of a specific module.",
          "parameters": [
            {
              "type": "integer",
              "description": "Module identifier",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/modules.ModuleResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            }
          }
        },
        "put": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "🍆 Module"
          ],
          "summary": "Update the source of a specific module.",
          "parameters": [
            {
              "type": "integer",
              "description": "Module identifier",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Module source",
              "name": "source",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/modules.ModuleResponse"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
      "errors.AppError": {
        "type": "object",
        "properties": {
          "err": {},
          "type": {
            "type": "string"
          }
        }
      },
...
      "modules.ModuleResponse": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}